Added fixes for End Project

+++ Projekt1/betterMatrixLib/include/matrixLib.h
@@ -21,15 +21,14 @@ void checkInput();
  * @param W - ilosc wierszy macierzy
  * @param K - ilosc kolumn w macierzy
  */
-void printMatrix(float **wynik, int W, int K);
-
+void printMatrix(double **wynik, int W, int K);
 /**
  * printMatrix - wyświtela wskazaną macierz
  * @param wynik -tablica, w ktorej zawarta jest macierz
  * @param W - ilosc wierszy macierzy
  * @param K - ilosc kolumn w macierzy
  */
-void printMatrix(long int **wynik, int W, int K);
+void printMatrix(int **wynik, int W, int K);
 
 /**
  * matrixAdd - dodaje 2 macierze
@@ -39,8 +38,7 @@ void printMatrix(long int **wynik, int W, int K);
  * @param K - ilosc kolumn w macierzy
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wynikowej
  */
-long int **matrixAdd(long int **a, long int **b, int W, int K);
-
+int **matrixAdd(int **a, int **b, int W, int K);
 /**
  * matrixAdd - dodaje 2 macierze
  * @param a - tablica, w ktorej zawarta jest macierz a
@@ -49,7 +47,7 @@ long int **matrixAdd(long int **a, long int **b, int W, int K);
  * @param K - ilosc kolumn w macierzy
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wynikowej
  */
-float **matrixAdd(float **a, float **b, int W, int K);
+double **matrixAdd(double **a, double **b, int W, int K);
 
 /**
  * matrixSubtract - odejmuje macierze
@@ -59,8 +57,7 @@ float **matrixAdd(float **a, float **b, int W, int K);
  * @param K - ilosc kolumn w macierzy
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-long int **matrixSubtract(long int **a, long int **b, int W, int K);
-
+int **matrixSubtract(int **a, int **b, int W, int K);
 /**
  * matrixSubtract - odejmuje macierze
  * @param a - tablica, w ktorej zawarta jest macierz a
@@ -69,7 +66,7 @@ long int **matrixSubtract(long int **a, long int **b, int W, int K);
  * @param K - ilosc kolumn w macierzy
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-float **matrixSubtract(float **a, float **b, int W, int K);
+double **matrixSubtract(double **a, double **b, int W, int K);
 
 /**
  * matrixMultiply - mnozy macierze
@@ -80,8 +77,7 @@ float **matrixSubtract(float **a, float **b, int W, int K);
  * @param Kb - ilosc kolumn macierzy b
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-long int **matrixMultiply(long int **a, long int **b, int Wa, int Ka, int Kb);
-
+int **matrixMultiply (int **a, int **b, int Wa, int Ka, int Kb);
 /**
  * matrixMultiply - mnozy macierze
  * @param a - tablica, w ktorej zawarta jest macierz a
@@ -91,7 +87,7 @@ long int **matrixMultiply(long int **a, long int **b, int Wa, int Ka, int Kb);
  * @param Kb - ilosc kolumn macierzy b
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-float **matrixMultiply(float **a, float **b, int Wa, int Ka, int Kb);
+double **matrixMultiply (double **a, double **b, int Wa, int Ka, int Kb);
 
 /**
  * matrixMultiplyByScalar -mnorzy macierz przez skalar
@@ -101,8 +97,7 @@ float **matrixMultiply(float **a, float **b, int Wa, int Ka, int Kb);
  * @param s - skalar
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-long int **matrixMultiplyByScalar(long int **a, int Wa, int Ka, long int s);
-
+int **matrixMultiplyByScalar (int **a, int Wa, int Ka, double s);
 /**
  * matrixMultiplyByScalar -mnorzy macierz przez skalar
  * @param a - tablica, w ktorej zawarta jest macierz a
@@ -111,7 +106,7 @@ long int **matrixMultiplyByScalar(long int **a, int Wa, int Ka, long int s);
  * @param s - skalar
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-float **matrixMultiplyByScalar(float **a, int Wa, int Ka, float s);
+double **matrixMultiplyByScalar (double **a, int Wa, int Ka, double s);
 
 /**
  * matrixTranspoze - transponuje macierz
@@ -120,8 +115,7 @@ float **matrixMultiplyByScalar(float **a, int Wa, int Ka, float s);
  * @param Ka -ilosc kolumn w macierzy
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-long int **matrixTranspoze(long int **a, int Wa, int Ka);
-
+int **matrixTranspoze (int **a, int Wa, int Ka);
 /**
  * matrixTranspoze - transponuje macierz
  * @param a - tablica, w ktorej zawarta jest macierz a
@@ -129,7 +123,7 @@ long int **matrixTranspoze(long int **a, int Wa, int Ka);
  * @param Ka -ilosc kolumn w macierzy
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-float **matrixTranspoze(float **a, int Wa, int Ka);
+double **matrixTranspoze  (double **a, int Wa, int Ka);
 
 /**
  * matrixPower - podnosi macierz do wskazanej potegi
@@ -139,8 +133,7 @@ float **matrixTranspoze(float **a, int Wa, int Ka);
  * @param st - stopien potegi
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-long int **matrixPower(long int **a, int Wa, int Ka, unsigned short st);
-
+int **matrixPower (int **a, int Wa, int Ka, unsigned st);
 /**
  * matrixPower - podnosi macierz do wskazanej potegi
  * @param a - tablica, w ktorej zawarta jest macierz a
@@ -149,15 +142,14 @@ long int **matrixPower(long int **a, int Wa, int Ka, unsigned short st);
  * @param st - stopien potegi
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-float **matrixPower(float **a, int Wa, int Ka, unsigned short st);
+double **matrixPower (double **a, int Wa, int Ka, unsigned st);
 
 /**
  * LUdecomposition -rozklad LU macierzy (do wyznacznika)
  * @param a -macierz do rozkladu
  * @param n -rozmiar macierzy (ilosc wierszy/kolumn - macierz musi byc kwadratowa)
  */
-bool LUdecomposition(int n, float **a);
-
+bool LUdecomposition(int n, double **a);
 /**
  * matrixDeterminant - liczy wyznacznik macierzy
  * @param a - tablica, w ktorej zawarta jest macierz a (jest to tablica typu double ze wzgledu na dzielenie w rozkladzie LU)
@@ -165,7 +157,7 @@ bool LUdecomposition(int n, float **a);
  * @param Ka -ilosc kolumn w macierzy
  * @return - zwraca wyznacznik macierzy
  */
-float matrixDeterminant(float **a, int Wa, int Ka);
+double matrixDeterminant (double **a, int Wa, int Ka);
 //tylko double ze wzgledu na dekompozycje lu gdzie jest dzielenie
 
 /**
@@ -175,8 +167,7 @@ float matrixDeterminant(float **a, int Wa, int Ka);
  * @param Ka -ilosc kolumn w macierzy
  * @return -zwraca watrosc "true" jesli macierz jest diagonalna lub wartosc "false" gdy nie jest diagonalna
  */
-bool matrixIsDiagonal(long int **a, int Wa, int Ka);
-
+bool matrixIsDiagonal(int **a, int Wa, int Ka);
 /**
  * matrixIsDiagonal - sprawdza czy macierz jest diagonalna
  * @param a -sprawdzana macierz
@@ -184,7 +175,7 @@ bool matrixIsDiagonal(long int **a, int Wa, int Ka);
  * @param Ka -ilosc kolumn w macierzy
  * @return -zwraca watrosc "true" jesli macierz jest diagonalna lub wartosc "false" gdy nie jest diagonalna
  */
-bool matrixIsDiagonal(float **a, int Wa, int Ka);
+bool matrixIsDiagonal(double **a, int Wa, int Ka);
 
 
 /**
@@ -192,14 +183,13 @@ bool matrixIsDiagonal(float **a, int Wa, int Ka);
  * @param a
  * @param b
  */
-void swap(long int *a, long int *b);
-
+void swap(int *a, int *b);
 /**
  * swap - zamienia dwie wartosci zmiennych miejscami
  * @param a
  * @param b
  */
-void swap(float *a, float *b);
+void swap(double *a, double *b);
 
 /**
  * sortRow - sortuje pierwszy rzad w macierzy (rosnaco)
@@ -207,15 +197,14 @@ void swap(float *a, float *b);
  * @param Ka -ilosc kolumn w macierzy
  * @return - zwraca wskaznik do podanej marcierzy z posortowanym pierwszym rzedem macierzy
  */
-long int **sortRow(long int **a, int Ka);
-
+int **sortRow(int **a, int Ka);
 /**
  * sortRow - sortuje pierwszy rzad w macierzy (rosnaco)
  * @param a - tablica, w ktorej zawarta jest macierz a
  * @param Ka -ilosc kolumn w macierzy
  * @return - zwraca wskaznik do podanej marcierzy z posortowanym pierwszym rzedem macierzy
  */
-float **sortRow(float **a, int Ka);
+double **sortRow(double **a, int Ka);
 
 /**
  * matrixSortRows -sortuje wszystkie rzedy macierzy (rosnaco)
@@ -224,8 +213,7 @@ float **sortRow(float **a, int Ka);
  * @param Ka -ilosc kolumn w macierzy
  * @return - zwraca wskaznik do marcierzy z posortowanymi rzedami
  */
-long int **matrixSortRows(long int **a, int Wa, int Ka);
-
+int **matrixSortRows (int **a, int Wa, int Ka);
 /**
  * matrixSortRows -sortuje wszystkie rzedy macierzy (rosnaco)
  * @param a - tablica, w ktorej zawarta jest macierz a
@@ -233,7 +221,7 @@ long int **matrixSortRows(long int **a, int Wa, int Ka);
  * @param Ka -ilosc kolumn w macierzy
  * @return - zwraca wskaznik do marcierzy z posortowanymi rzedami
  */
-float **matrixSortRows(float **a, int Wa, int Ka);
+double **matrixSortRows (double **a, int Wa, int Ka);
 
 //do zrobienia
 /**
@@ -244,8 +232,7 @@ float **matrixSortRows(float **a, int Wa, int Ka);
  * @param max - maksymalna wartosc do ktorej losowane beda liczby
  * @return - zwraca wskaznik do losowo wygenerowanej macierzy
  */
-long int **matrixGenerate(int W, int K, long int min, long int max);
-
+int **matrixGenerate(int W, int K,int min,int max);
 /**
  * matrixGenerate - generuje losowa macierz
  * @param Wa - ilosc wierszy w macierzy
@@ -254,6 +241,5 @@ long int **matrixGenerate(int W, int K, long int min, long int max);
  * @param max - maksymalna wartosc do ktorej losowane beda liczby
  * @return - zwraca wskaznik do losowo wygenerowanej macierzy
  */
-float **matrixGenerate(int W, int K, float min, float max);
-
+double **matrixGenerate (int W, int K, double min, double max);
 #endif //PROJEKT1_MATRIXLIB_H
